Added tempo buffer screen upload and download | Added popup for edit i.e point, polygon, line (to be fixed) | Added polygon save function | Added Edit Save Function | Workaround for page refresh now avaialble for layer_ID only

+++ frontend/Component/List.tsx
@@ -1,6 +1,10 @@
 import axios from "axios";
-import { useEffect } from "react";
-import type { FeatureCollection } from "geojson";
+import React from "react";
+import { useEffect, useState } from "react";
+import tokml from "geojson-to-kml";
+import type { Feature, FeatureCollection, Geometry } from "geojson";
+import JSZip from "jszip";
+import type { UploadStatus } from "./Floating";
 
 type KmzFile = {
   id: string;
@@ -21,6 +25,7 @@ type ListProps = {
   layerList: KmzFile[] | null;
   openLayer: string | null;
   exportFC: FeatureCollection;
+  setStatus: (a : UploadStatus) => void;
 };
 
 export const List = ({
@@ -31,10 +36,13 @@ export const List = ({
   layerList,
   openLayer,
   exportFC,
+  setStatus
 }: ListProps) => {
   // Impossible To Keep Styles (View Only)
   async function downloadKMZ(fileId: string, fileName: string) {
     try {
+      setStatus("downloading");
+
       const response = await axios.get(
         `http://localhost:3000/download/${fileId}`,
         {
@@ -51,29 +59,15 @@ export const List = ({
       link.click();
 
       link.remove();
+      
+
       window.URL.revokeObjectURL(url);
+      setTimeout(() => { setStatus("idle")}, 2000); // tempo
+
     } catch (error) {
       console.error("Download failed:", error);
     }
-    // const kml = tokml(fc as any, {
-    //   name: "name",
-    //   description: "description",
-    //   simplestyle: true,
-    // });
-    // const zip = new JSZip();
-    // zip.file("doc.kml", kml); // KMZ expects 'doc.kml'
-    // const kmzBlob = await zip.generateAsync({
-    //   type: "blob",
-    //   compression: "DEFLATE",
-    // });
-    //
-    // const a = document.createElement("a");
-    // a.href = URL.createObjectURL(
-    //   new Blob([kmzBlob], { type: "application/vnd.google-earth.kmz" }),
-    // );
-    // a.download = `${name}(Copy).kmz`;
-    // a.click();
-    // URL.revokeObjectURL(a.href);
+    
   }
 
   function shortenName(name: string, maxLength = 25) {
@@ -84,9 +78,22 @@ export const List = ({
   }
 
   async function handleDelete(fileId: string) {
+
+    if (!confirm("Delete this file?")) return;
+
     try {
       await axios.delete(`http://localhost:3000/files/${fileId}`);
+
+      const wasOpen = openLayer === fileId;
+
+
       getKMZList();
+
+      window.dispatchEvent(
+        new CustomEvent("kmz:changed", {
+          detail: { fileId: wasOpen ? null : openLayer }
+        })
+      );
     } catch (error) {
       console.log(error);
     }
